Enable offline saving for raygun
In the main.ts file, an additional configuration option for Raygun was added. The method 'saveIfOffline' was set to true which allows for error logs to be saved and sent when the app is back online if it was offline when the error occurred. This ensures no error information is lost due to poor network conditions.