Add a second error trigger and error reporting handler
Added a new button `Trigger error two` that triggers a different error method. This new method called `triggerErrorTwo` throws a specified Error and catches it to send the error information with `rg4js`. This allows for the error to be manually logged, providing more control and visibility to the type and handling of errors in the application.